+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - `cli/flow`: Add workflow subcommands
 - `cli/sync`: Add option not to switch to directory
+- `cli/add`: Add option not to switch to directory
 
 ### Fixed